+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 
 ---
+## [0.8.0] - 2022-06-23
+### Added
+- [pyo3] ABE key delegation
+- [pyo3 + JS/bindgen + FFI] Add attributes rotation mechanism
+### Changed
+- In struct `Policy`:
+  * `last_attribute` becomes `last_attribute_value`
+  * `max_attribute` becomes `max_attribute_value`
+- Add to Python Notebook Delegation mechanism
+### Fixed
+### Removed
+---
+
+---
 ## [0.7.0] - 2022-06-16
 ### Added
 - [pyo3] Add Rust bindings for Python thanks to pyo3